# postman/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Govia API Docs",
    "version": "1.0.0",
    "description": "Comprehensive REST API & Real-Time Socket documentation for the Govia platform."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API Version 1 Base Path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT Bearer token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "LoginResponseData": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsIn..."
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsIn..."
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ],
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "User logged in successfully"
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number",
                "example": 1
              },
              "limit": {
                "type": "number",
                "example": 10
              },
              "totalPage": {
                "type": "number",
                "example": 1
              },
              "total": {
                "type": "number",
                "example": 1
              }
            },
            "required": [
              "page",
              "limit",
              "totalPage",
              "total"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/LoginResponseData"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "MessageResponseData": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "65fe1234567890abcdef1234"
          },
          "conversationId": {
            "type": "string",
            "example": "651234567890abcdef123456"
          },
          "sender": {
            "nullable": true
          },
          "receiver": {
            "nullable": true
          },
          "text": {
            "type": "string"
          },
          "messageType": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "file",
              "meeting"
            ],
            "example": "text"
          },
          "meetingId": {
            "nullable": true
          },
          "attachment": {
            "type": "string"
          },
          "read": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "conversationId",
          "messageType",
          "read",
          "createdAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/message": {
      "post": {
        "summary": "Send Message",
        "description": "Sends a text message or file attachment to a conversation.",
        "tags": [
          "Message"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "conversationId": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  },
                  "messageType": {
                    "type": "string",
                    "enum": [
                      "text",
                      "image",
                      "file",
                      "meeting"
                    ]
                  },
                  "meetingId": {
                    "type": "string"
                  }
                },
                "required": [
                  "conversationId"
                ],
                "description": "Send chat message payload",
                "example": {
                  "conversationId": "651234567890abcdef123456",
                  "text": "Hello, I have a question about my consultation.",
                  "messageType": "text"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string",
                          "example": "Operation successful"
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "number",
                              "example": 1
                            },
                            "limit": {
                              "type": "number",
                              "example": 10
                            },
                            "totalPage": {
                              "type": "number",
                              "example": 1
                            },
                            "total": {
                              "type": "number",
                              "example": 1
                            }
                          },
                          "required": [
                            "page",
                            "limit",
                            "totalPage",
                            "total"
                          ]
                        },
                        "data": {
                          "$ref": "#/components/schemas/MessageResponseData"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/message/meeting": {
      "post": {
        "summary": "Create Meeting In Chat (\"Now\" or \"Later\")",
        "description": "Creates an instant or scheduled Zoom meeting directly within a conversation, generating a chat card and syncing to Schedule Page.",
        "tags": [
          "Message"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "conversationId": {
                    "type": "string"
                  },
                  "topic": {
                    "type": "string"
                  },
                  "meetingType": {
                    "type": "string",
                    "enum": [
                      "INSTANT",
                      "SCHEDULED"
                    ]
                  },
                  "startTime": {
                    "type": "string"
                  },
                  "durationMinutes": {
                    "type": "number",
                    "minimum": 5,
                    "maximum": 1440
                  },
                  "timezone": {
                    "type": "string"
                  },
                  "agenda": {
                    "type": "string"
                  }
                },
                "required": [
                  "conversationId",
                  "topic",
                  "meetingType"
                ],
                "description": "Create instant or scheduled meeting from inside chat thread",
                "example": {
                  "conversationId": "651234567890abcdef123456",
                  "topic": "Case Discussion",
                  "meetingType": "SCHEDULED",
                  "startTime": "2026-09-15T15:00:00.000Z",
                  "durationMinutes": 30
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Meeting created and posted to chat",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string",
                          "example": "Operation successful"
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "number",
                              "example": 1
                            },
                            "limit": {
                              "type": "number",
                              "example": 10
                            },
                            "totalPage": {
                              "type": "number",
                              "example": 1
                            },
                            "total": {
                              "type": "number",
                              "example": 1
                            }
                          },
                          "required": [
                            "page",
                            "limit",
                            "totalPage",
                            "total"
                          ]
                        },
                        "data": {
                          "nullable": true
                        }
                      },
                      "required": [
                        "success",
                        "statusCode"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/message/search-users": {
      "get": {
        "summary": "Search Users Across All Roles",
        "description": "Searches users by name, role, email with precomputed direct conversation IDs for 1-click messaging.",
        "tags": [
          "Message"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "searchTerm",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved for messaging",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string",
                          "example": "Operation successful"
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "number",
                              "example": 1
                            },
                            "limit": {
                              "type": "number",
                              "example": 10
                            },
                            "totalPage": {
                              "type": "number",
                              "example": 1
                            },
                            "total": {
                              "type": "number",
                              "example": 1
                            }
                          },
                          "required": [
                            "page",
                            "limit",
                            "totalPage",
                            "total"
                          ]
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "nullable": true
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/message/open-chat": {
      "post": {
        "summary": "Open or Create Chat Thread",
        "description": "Click-action on search result to retrieve existing conversation or create new thread atomically.",
        "tags": [
          "Message"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "participantId": {
                    "type": "string"
                  }
                },
                "required": [
                  "participantId"
                ],
                "description": "Click-to-chat payload with target user ID",
                "example": {
                  "participantId": "64bcde1234567890abcdef12"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conversation retrieved or created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string",
                          "example": "Operation successful"
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "number",
                              "example": 1
                            },
                            "limit": {
                              "type": "number",
                              "example": 10
                            },
                            "totalPage": {
                              "type": "number",
                              "example": 1
                            },
                            "total": {
                              "type": "number",
                              "example": 1
                            }
                          },
                          "required": [
                            "page",
                            "limit",
                            "totalPage",
                            "total"
                          ]
                        },
                        "data": {
                          "nullable": true
                        }
                      },
                      "required": [
                        "success",
                        "statusCode"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/message/{conversationId}": {
      "get": {
        "summary": "Get Messages by Conversation",
        "description": "Retrieves chronological chat history with populated meeting status and attachments.",
        "tags": [
          "Message"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string",
                          "example": "Operation successful"
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "number",
                              "example": 1
                            },
                            "limit": {
                              "type": "number",
                              "example": 10
                            },
                            "totalPage": {
                              "type": "number",
                              "example": 1
                            },
                            "total": {
                              "type": "number",
                              "example": 1
                            }
                          },
                          "required": [
                            "page",
                            "limit",
                            "totalPage",
                            "total"
                          ]
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MessageResponseData"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/message/read/{conversationId}": {
      "patch": {
        "summary": "Mark Messages as Read",
        "description": "Marks unread messages in the thread as read and emits real-time read receipt.",
        "tags": [
          "Message"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages marked as read",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string",
                          "example": "Operation successful"
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "number",
                              "example": 1
                            },
                            "limit": {
                              "type": "number",
                              "example": 10
                            },
                            "totalPage": {
                              "type": "number",
                              "example": 1
                            },
                            "total": {
                              "type": "number",
                              "example": 1
                            }
                          },
                          "required": [
                            "page",
                            "limit",
                            "totalPage",
                            "total"
                          ]
                        },
                        "data": {
                          "nullable": true
                        }
                      },
                      "required": [
                        "success",
                        "statusCode"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}